requirements.txt
django-compat
pymysql
